Move forum thread stuff to its own files
